Updated units in GLScene_RunTime_FMX.dpk
git-svn-id: http://svn.code.sf.net/p/glscene/code@6616 c452ef9c-9f09-11df-acb8-e9cdb1ccbceb

+++ branches/GLScene_FMX/Source/GLS.Scene.pas
@@ -2071,7 +2071,7 @@ type
        are further from the eye, than its centre.}
     function PixelToDistance(x, y: integer): Single;
     {: Design time notification }
-    procedure NotifyMouseMove(Shift: TShiftState; X, Y: Integer);
+    procedure NotifyMouseMove(Shift: TShiftState; X, Y: Single);
 
     {: Renders the scene on the viewer.<p>
        You do not need to call this method, unless you explicitly want a
@@ -9249,7 +9249,7 @@ end;
 // NotifyMouseMove
 //
 
-procedure TGLSceneBuffer.NotifyMouseMove(Shift: TShiftState; X, Y: Integer);
+procedure TGLSceneBuffer.NotifyMouseMove(Shift: TShiftState; X, Y: Single);
 begin
   // Nothing
 end;
@@ -10214,3 +10214,4 @@ initialization
   QueryPerformanceFrequency(vCounterFrequency);
 
 end.
+
